Move providerEngine into it's own file
