+++ src/scenarios/all.ts
@@ -1,13 +1,13 @@
-import { providerEngine } from '../contracts';
+import { providerEngine } from '../provider_engine';
 
-import { scenario as cancelOrdersUpTo } from './cancel_orders_up_to';
-import { scenario as executeTransaction } from './execute_transaction';
-import { scenario as fillOrderERC20 } from './fill_order_erc20';
-import { scenario as fillOrderERC721 } from './fill_order_erc721';
-import { scenario as fillOrderFees } from './fill_order_fees';
-import { scenario as forwarder_buy_erc20_tokens } from './forwarder_buy_erc20_tokens';
-import { scenario as forwarder_buy_erc721_tokens } from './forwarder_buy_erc721_tokens';
-import { scenario as matchOrders } from './match_orders';
+import { scenarioAsync as cancelOrdersUpTo } from './cancel_orders_up_to';
+import { scenarioAsync as executeTransaction } from './execute_transaction';
+import { scenarioAsync as fillOrderERC20 } from './fill_order_erc20';
+import { scenarioAsync as fillOrderERC721 } from './fill_order_erc721';
+import { scenarioAsync as fillOrderFees } from './fill_order_fees';
+import { scenarioAsync as forwarder_buy_erc20_tokens } from './forwarder_buy_erc20_tokens';
+import { scenarioAsync as forwarder_buy_erc721_tokens } from './forwarder_buy_erc721_tokens';
+import { scenarioAsync as matchOrders } from './match_orders';
 
 void (async () => {
     try {
